[Feat] Inventory Packet
장비 인벤토리와 소비 인벤토리 기타 인벤토리를 각 구성하여 해결

+++ Source/RLR/GameManager/RLRStruct.h
@@ -244,13 +244,14 @@ struct FItemData : public FTableRowBase
 
 	//DB에 저장되어야할 정보
 	int32 ITEM_SLOT_IDX;
-	int32 ITEM_ID;
+	int64 ITEM_ID;
 	bool IsEquiped;
 
 	//나중에 패킷 날라오면, 그 정보로 FItemData를 만들어준다.
 	void MakeItemData(const Protocol::Item itemData);
 
 	void SetItemSlotIndex(int32 Id){ITEM_SLOT_IDX = Id;}
+
 	
 };
